tests/functional/requirements.txt
aiohttp==3.8.6
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1

redis==4.4.2
elasticsearch==8.10.1
